refactor(api): move campaign participant activity to ist bounded context
