+++ api/src/prescription/campaign/application/campaign-detail-route.js
@@ -2,6 +2,11 @@ import Joi from 'joi';
 import { campaignDetailController } from './campaign-detail-controller.js';
 import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
+
+import { CampaignParticipationStatuses } from '../../shared/domain/constants.js';
+
+const campaignParticipationStatuses = Object.values(CampaignParticipationStatuses);
+
 const register = async function (server) {
   server.route([
     {
@@ -71,6 +76,33 @@ const register = async function (server) {
           "- **Cette route est restreinte via un token dédié passé en paramètre avec l'id de l'utilisateur.**\n" +
             "- Récupération d'un CSV avec les résultats de la campagne d‘évaluation\n" +
             '- L‘utilisateur doit avoir les droits d‘accès à l‘organisation liée à la campagne à créer',
+        ],
+        tags: ['api', 'campaign'],
+      },
+    },
+    {
+      method: 'GET',
+      path: '/api/campaigns/{id}/participants-activity',
+      config: {
+        validate: {
+          params: Joi.object({
+            id: identifiersType.campaignId,
+          }),
+          query: Joi.object({
+            'page[number]': Joi.number().integer().empty(''),
+            'page[size]': Joi.number().integer().empty(''),
+            'filter[divisions][]': [Joi.string(), Joi.array().items(Joi.string())],
+            'filter[status]': Joi.string()
+              .valid(...campaignParticipationStatuses)
+              .empty(''),
+            'filter[groups][]': [Joi.string(), Joi.array().items(Joi.string())],
+            'filter[search]': Joi.string().empty(''),
+          }),
+        },
+        handler: campaignDetailController.findParticipantsActivity,
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            "- Récupération des participations d'une campagne par son id",
         ],
         tags: ['api', 'campaign'],
       },
