+++ api/src/prescription/campaign/domain/read-models/CampaignParticipantActivity.js
