+++ api/src/prescription/campaign/domain/usecases/find-paginated-campaign-participants-activities.js
@@ -1,4 +1,4 @@
-import { UserNotAuthorizedToAccessEntityError } from '../errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../lib/domain/errors.js';
 
 const findPaginatedCampaignParticipantsActivities = async function ({
   userId,
