+++ api/src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-participant-activity-serializer.js
