+++ api/tests/prescription/campaign/integration/infrastructure/repositories/campaign-participant-activity-repository_test.js
@@ -1,6 +1,6 @@
-import { expect, databaseBuilder } from '../../../test-helper.js';
-import { campaignParticipantActivityRepository } from '../../../../lib/infrastructure/repositories/campaign-participant-activity-repository.js';
-import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
+import { expect, databaseBuilder } from '../../../../../test-helper.js';
+import { campaignParticipantActivityRepository } from '../../../../../../src/prescription/campaign/infrastructure/repositories/campaign-participant-activity-repository.js';
+import { CampaignParticipationStatuses } from '../../../../../../src/prescription/shared/domain/constants.js';
 
 const { STARTED, SHARED, TO_SHARE } = CampaignParticipationStatuses;
