+++ api/tests/prescription/campaign/unit/application/campaign-detail-controller_test.js
@@ -271,6 +271,56 @@ describe('Unit | Application | Controller | Campaign detail', function () {
       );
     });
   });
+
+  describe('#findParticipantsActivity', function () {
+    let serializedParticipantsActivities;
+    let participantsActivities;
+    const filters = { status: 'SHARED', groups: ['L1'], search: 'Choupette' };
+
+    const campaignId = 1;
+    const userId = 1;
+    let campaignParticipantsActivitySerializerStub;
+    beforeEach(function () {
+      participantsActivities = Symbol('participants activities');
+      serializedParticipantsActivities = Symbol('serialized participants activities');
+      sinon.stub(usecases, 'findPaginatedCampaignParticipantsActivities');
+      campaignParticipantsActivitySerializerStub = {
+        serialize: sinon.stub(),
+      };
+    });
+
+    it('should return the participants activities properly serialized', async function () {
+      // given
+      usecases.findPaginatedCampaignParticipantsActivities
+        .withArgs({ campaignId, userId, page: { number: 3 }, filters })
+        .resolves(participantsActivities);
+      campaignParticipantsActivitySerializerStub.serialize
+        .withArgs(participantsActivities)
+        .returns(serializedParticipantsActivities);
+
+      // when
+      const response = await campaignDetailController.findParticipantsActivity(
+        {
+          params: { id: campaignId },
+          auth: {
+            credentials: { userId },
+          },
+
+          query: {
+            'page[number]': 3,
+            'filter[groups][]': ['L1'],
+            'filter[status]': 'SHARED',
+            'filter[search]': 'Choupette',
+          },
+        },
+        hFake,
+        { campaignParticipantsActivitySerializer: campaignParticipantsActivitySerializerStub },
+      );
+
+      // then
+      expect(response).to.be.equal(serializedParticipantsActivities);
+    });
+  });
 });
 
 function _getRequestForCampaignId(campaignId) {
