+++ api/tests/prescription/campaign/unit/infrastrucutre/serializers/jsonapi/campaign-participant-activity-serializer_test.js
@@ -1,7 +1,7 @@
-import { expect } from '../../../../test-helper.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/campaign-participant-activity-serializer.js';
-import { CampaignParticipantActivity } from '../../../../../lib/domain/read-models/CampaignParticipantActivity.js';
-import { CampaignParticipationStatuses } from '../../../../../src/prescription/shared/domain/constants.js';
+import { expect } from '../../../../../../test-helper.js';
+import * as serializer from './../../../../../../../src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-participant-activity-serializer.js';
+import { CampaignParticipantActivity } from '../../../../../../../src/prescription/campaign/domain/read-models/CampaignParticipantActivity.js';
+import { CampaignParticipationStatuses } from '../../../../../../../src/prescription/shared/domain/constants.js';
 
 const { SHARED, STARTED } = CampaignParticipationStatuses;
