+++ api/tests/unit/application/campaign/campaign-controller_test.js
@@ -242,54 +242,4 @@ describe('Unit | Application | Controller | Campaign', function () {
       expect(response).to.be.equal(serializedCampaign);
     });
   });
-
-  describe('#findParticipantsActivity', function () {
-    let serializedParticipantsActivities;
-    let participantsActivities;
-    const filters = { status: 'SHARED', groups: ['L1'], search: 'Choupette' };
-
-    const campaignId = 1;
-    const userId = 1;
-    let campaignParticipantsActivitySerializerStub;
-    beforeEach(function () {
-      participantsActivities = Symbol('participants activities');
-      serializedParticipantsActivities = Symbol('serialized participants activities');
-      sinon.stub(usecases, 'findPaginatedCampaignParticipantsActivities');
-      campaignParticipantsActivitySerializerStub = {
-        serialize: sinon.stub(),
-      };
-    });
-
-    it('should return the participants activities properly serialized', async function () {
-      // given
-      usecases.findPaginatedCampaignParticipantsActivities
-        .withArgs({ campaignId, userId, page: { number: 3 }, filters })
-        .resolves(participantsActivities);
-      campaignParticipantsActivitySerializerStub.serialize
-        .withArgs(participantsActivities)
-        .returns(serializedParticipantsActivities);
-
-      // when
-      const response = await campaignController.findParticipantsActivity(
-        {
-          params: { id: campaignId },
-          auth: {
-            credentials: { userId },
-          },
-
-          query: {
-            'page[number]': 3,
-            'filter[groups][]': ['L1'],
-            'filter[status]': 'SHARED',
-            'filter[search]': 'Choupette',
-          },
-        },
-        hFake,
-        { campaignParticipantsActivitySerializer: campaignParticipantsActivitySerializerStub },
-      );
-
-      // then
-      expect(response).to.be.equal(serializedParticipantsActivities);
-    });
-  });
 });
